big update

+++ database/migrations/2024_05_15_021232_create_businesses_table.php
@@ -17,7 +17,7 @@ class CreateBusinessesTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('phone');
-            $table->integer('fee_percent');
+            $table->float('fee_percent');
             $table->string('card_number');
             $table->bigInteger('total_money');
             $table->enum('formality', ['R', 'D']);
